Move redirect streaming callback test to simple_httpclient only.
This test passes with recent versions of libcurl but fails
with the version available on travis-ci.
